Extract types tests
